requirements.txt
fastapi
tweepy
python-dotenv
google-generativeai
Pillow
uvicorn
